pyproject.toml
[project]
name = "inside-out"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gtts>=2.5.4",
    "librosa>=0.11.0",
    "noisereduce>=3.0.3",
    "numpy>=2.2.5",
    "openai-whisper>=20240930",
    "pysrt>=1.1.2",
    "scipy>=1.15.2",
    "torch>=2.2.2",
    "torchaudio>=2.2.2",      
    "torchvision>=0.17.2",     
    "tqdm>=4.66.4",            
    "scikit-learn>=1.4.2",     
    "matplotlib>=3.8.4",       
    "moviepy>=1.0.3",         
    "moviepy<=1.9.9",
]
